+++ app/Http/Controllers/ContactPersonController.php
@@ -11,6 +11,8 @@ class ContactPersonController extends Controller
 {
     public function index(Customer $customer)
     {
+        $this->authorize('viewAny', ContactPerson::class);
+
         $contactpersons = ContactPerson::where('customer_id', $customer->id)->get();
 
         return view('contactperson.index', compact('customer', 'contactpersons'));
@@ -18,11 +20,15 @@ class ContactPersonController extends Controller
 
     public function create(Customer $customer)
     {
+        $this->authorize('create', ContactPerson::class);
+
         return view('contactperson.create', compact('customer'));
     }
 
     public function store(Customer $customer, ContactPersonRequest $request)
     {
+        $this->authorize('create', ContactPerson::class);
+
         $customer->contactpersons()->create($request->validated());
 
         return redirect(route('contactperson.index', $customer));
@@ -30,11 +36,15 @@ class ContactPersonController extends Controller
 
     public function edit(Customer $customer, ContactPerson $contactperson)
     {
+        $this->authorize('update', ContactPerson::class);
+
         return view('contactperson.edit', compact('customer', 'contactperson'));
     }
 
     public function update(Customer $customer, ContactPerson $contactperson, ContactPersonRequest $request)
     {
+        $this->authorize('update', ContactPerson::class);
+
         $contactperson->update($request->validated());
 
         return redirect(route('contactperson.index', $customer));
@@ -42,6 +52,8 @@ class ContactPersonController extends Controller
 
     public function destroy(Customer $customer, ContactPerson $contactperson)
     {
+        $this->authorize('delete', ContactPerson::class);
+
         $contactperson->delete();
 
         return redirect(route('contactperson.index', $customer));
