+++ app/Http/Controllers/DECTController.php
@@ -11,6 +11,8 @@ class DECTController extends Controller
 {
     public function index(Customer $customer)
     {
+        $this->authorize('viewAny', DECT::class);
+
         $dect = $this->getFilteredQuery(DECT::class, $customer)
                        ->get();
 
@@ -19,6 +21,8 @@ class DECTController extends Controller
 
     public function create(Customer $customer)
     {
+        $this->authorize('create', DECT::class);
+
         $sites = $this->getSitesForCustomer($customer);
 
         return view('dect.create', compact('customer', 'sites'));
@@ -26,6 +30,8 @@ class DECTController extends Controller
 
     public function store(Customer $customer, DECTRequest $request)
     {
+        $this->authorize('create', DECT::class);
+
         $customer->dect()->create($request->validated());
 
         return redirect(route('dect.index', $customer));
@@ -33,6 +39,8 @@ class DECTController extends Controller
 
     public function edit(Customer $customer, DECT $dect)
     {
+        $this->authorize('update', DECT::class);
+
         $sites = $this->getSitesForCustomer($customer);
 
         return view('dect.edit', compact('customer', 'sites', 'dect'));
@@ -40,6 +48,8 @@ class DECTController extends Controller
 
     public function update(Customer $customer, DECT $dect, DECTRequest $request)
     {
+        $this->authorize('update', DECT::class);
+
         $dect->update($request->validated());
 
         return redirect(route('dect.index', $customer));
@@ -47,6 +57,8 @@ class DECTController extends Controller
 
     public function destroy(Customer $customer, DECT $dect)
     {
+        $this->authorize('delete', DECT::class);
+
         $dect->delete();
 
         return redirect(route('dect.index', $customer));
