+++ app/Http/Controllers/LicenseSoftwareController.php
@@ -11,16 +11,22 @@ class LicenseSoftwareController extends Controller
 {
     public function index(Customer $customer)
     {
+        $this->authorize('viewAny', LicenseSoftware::class);
+
         return view('licensesoftware.index', compact('customer'));
     }
 
     public function create(Customer $customer)
     {
+        $this->authorize('create', LicenseSoftware::class);
+
         return view('licensesoftware.create', compact('customer'));
     }
 
     public function store(Customer $customer, LicenseSoftwareRequest $request)
     {
+        $this->authorize('create', LicenseSoftware::class);
+
         $customer->licensesoftware()->create($request->validated());
 
         return redirect(route('licensesoftware.index', $customer));
@@ -28,11 +34,15 @@ class LicenseSoftwareController extends Controller
 
     public function edit(Customer $customer, LicenseSoftware $licensesoftware)
     {
+        $this->authorize('update', LicenseSoftware::class);
+
         return view('licensesoftware.edit', compact('customer', 'licensesoftware'));
     }
 
     public function update(Customer $customer, LicenseSoftware $licensesoftware, LicenseSoftwareRequest $request)
     {
+        $this->authorize('update', LicenseSoftware::class);
+
         $licensesoftware->update($request->validated());
 
         return redirect(route('licensesoftware.index', $customer));
@@ -40,6 +50,8 @@ class LicenseSoftwareController extends Controller
 
     public function destroy(Customer $customer, LicenseSoftware $licensesoftware)
     {
+        $this->authorize('delete', LicenseSoftware::class);
+
         $licensesoftware->delete();
 
         return redirect(route('licensesoftware.index', $customer));
