+++ app/Http/Controllers/LicenseWindowsController.php
@@ -12,6 +12,8 @@ class LicenseWindowsController extends Controller
 {
     public function index(Customer $customer)
     {
+        $this->authorize('viewAny', LicenseWindows::class);
+
         $licensewindows = $this->getFilteredQuery(LicenseWindows::class, $customer)
                         ->with('operatingSystem')
                         ->get();
@@ -21,6 +23,8 @@ class LicenseWindowsController extends Controller
 
     public function create(Customer $customer)
     {
+        $this->authorize('create', LicenseWindows::class);
+
         $operatingSystems = OperatingSystem::all();
 
         return view('licensewindows.create', compact('customer', 'operatingSystems'));
@@ -28,6 +32,8 @@ class LicenseWindowsController extends Controller
 
     public function store(Customer $customer, LicenseWindowsRequest $request)
     {
+        $this->authorize('create', LicenseWindows::class);
+
         $customer->licensewindows()->create($request->validated());
 
         return redirect(route('licensewindows.index', $customer));
@@ -35,6 +41,8 @@ class LicenseWindowsController extends Controller
 
     public function edit(Customer $customer, LicenseWindows $licensewindows)
     {
+        $this->authorize('update', LicenseWindows::class);
+
         $operatingSystems = OperatingSystem::all();
 
         return view('licensewindows.edit', compact('customer', 'operatingSystems', 'licensewindows'));
@@ -42,6 +50,8 @@ class LicenseWindowsController extends Controller
 
     public function update(Customer $customer, LicenseWindows $licensewindows, LicenseWindowsRequest $request)
     {
+        $this->authorize('update', LicenseWindows::class);
+
         $licensewindows->update($request->validated());
 
         return redirect(route('licensewindows.index', $customer));
@@ -49,6 +59,8 @@ class LicenseWindowsController extends Controller
 
     public function destroy(Customer $customer, LicenseWindows $licensewindows)
     {
+        $this->authorize('delete', LicenseWindows::class);
+
         $licensewindows->delete();
 
         return redirect(route('licensewindows.index', $customer));
