+++ app/Http/Controllers/LoginWebsiteController.php
@@ -11,16 +11,22 @@ class LoginWebsiteController extends Controller
 
     public function index(Customer $customer)
     {
+        $this->authorize('viewAny', LoginWebsite::class);
+
         return view('loginwebsite.index', compact('customer'));
     }
 
     public function create(Customer $customer)
     {
+        $this->authorize('create', LoginWebsite::class);
+
         return view('loginwebsite.create', compact('customer'));
     }
 
     public function store(Customer $customer, LoginWebsiteRequest $request)
     {
+        $this->authorize('create', LoginWebsite::class);
+
         $customer->loginwebsites()->create($request->validated());
 
         return redirect(route('loginwebsite.index', $customer));
@@ -28,11 +34,15 @@ class LoginWebsiteController extends Controller
 
     public function edit(Customer $customer, LoginWebsite $loginwebsite)
     {
+        $this->authorize('update', LoginWebsite::class);
+
         return view('loginwebsite.edit', compact('customer', 'loginwebsite'));
     }
 
     public function update(Customer $customer, LoginWebsite $loginwebsite, LoginWebsiteRequest $request)
     {
+        $this->authorize('update', LoginWebsite::class);
+
         $loginwebsite->update($request->validated());
 
         return redirect(route('loginwebsite.index', $customer));
@@ -40,6 +50,8 @@ class LoginWebsiteController extends Controller
 
     public function destroy(Customer $customer, LoginWebsite $loginwebsite)
     {
+        $this->authorize('delete', LoginWebsite::class);
+
         $loginwebsite->delete();
 
         return redirect(route('loginwebsite.index', $customer));
