+++ app/Http/Controllers/PhoneController.php
@@ -10,6 +10,8 @@ class PhoneController extends Controller
 {
     public function index(Customer $customer)
     {
+        $this->authorize('viewAny', Phone::class);
+
         $phones = $this->getFilteredQuery(Phone::class, $customer)
                        ->get();
 
@@ -18,6 +20,8 @@ class PhoneController extends Controller
 
     public function create(Customer $customer)
     {
+        $this->authorize('create', Phone::class);
+
         $sites = $this->getSitesForCustomer($customer);
 
         return view('phone.create', compact('customer', 'sites'));
@@ -25,6 +29,8 @@ class PhoneController extends Controller
 
     public function store(Customer $customer, PhoneRequest $request)
     {
+        $this->authorize('create', Phone::class);
+
         $customer->phones()->create($request->validated());
 
         return redirect(route('phone.index', $customer));
@@ -32,6 +38,8 @@ class PhoneController extends Controller
 
     public function edit(Customer $customer, Phone $phone)
     {
+        $this->authorize('update', Phone::class);
+
         $sites = $this->getSitesForCustomer($customer);
 
         return view('phone.edit', compact('customer', 'sites', 'phone'));
@@ -39,6 +47,8 @@ class PhoneController extends Controller
 
     public function update(Customer $customer, Phone $phone, PhoneRequest $request)
     {
+        $this->authorize('update', Phone::class);
+
         $phone->update($request->validated());
 
         return redirect(route('phone.index', $customer));
@@ -46,6 +56,8 @@ class PhoneController extends Controller
 
     public function destroy(Customer $customer, Phone $phone)
     {
+        $this->authorize('delete', Phone::class);
+
         $phone->delete();
 
         return redirect(route('phone.index', $customer));
